setting :: Resources, Sources, Applications 나누기 & Scene, App Delegate 설정
