+++ ROOT/ROOT/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
