+++ ROOT/ROOT/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
