+++ ROOT/ROOT/Resources/Assets.xcassets/Contents.json
